requirements.txt
discord-py-interactions==5.15.0
psycopg==3.2.10
psycopg-pool==3.2.6
pydantic==2.11.7
pydantic-settings==2.10.1
pytest==8.4.2
pytest-asyncio==1.2.0
